# docs/v150-openapi.yaml
openapi: 3.0.3
info:
  title: 7zi API
  description: |
    # 7zi - AI 驱动的团队管理平台 API
    
    11 位 AI 成员 · 24/7 自主工作 · 实时协作
    
    ## 功能特性
    - 🤖 AI Agent 智能调度系统
    - 🔄 WebSocket 实时协作
    - 🔐 RBAC 权限控制
    - 📊 性能监控与告警
    - 🌍 多语言国际化
    
    ## 认证方式
    所有 API 请求需要在 Header 中携带 JWT Token:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## 错误响应格式
    所有错误响应遵循统一格式:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Error description",
        "details": {}
      }
    }
    ```
    
    ## 限流策略
    - 默认限制: 100 请求/分钟/IP
    - 认证用户: 200 请求/分钟/用户
    - API Key 用户: 500 请求/分钟
    
  version: 1.5.0
  contact:
    name: 7zi API Support
    email: api@7zi.com
    url: https://7zi.com/docs/api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://7zi.com/api
    description: 生产环境
  - url: http://localhost:3000/api
    description: 开发环境
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 认证令牌
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          default: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - VALIDATION_ERROR
                - INTERNAL_ERROR
                - RATE_LIMIT_EXCEEDED
                - DATABASE_ERROR
                - CONFLICT
            message:
              type: string
            details:
              type: object
...
    QueueStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            status:
              type: string
              enum: [ok, degraded, error]
            stats:
              type: object
              properties:
                total:
                  type: integer
                pending:
                  type: integer
                inProgress:
                  type: integer
                completed:
                  type: integer
                failed:
                  type: integer
                byPriority:
                  type: object
                  properties:
                    urgent:
                      type: integer
                    high:
                      type: integer
                    normal:
                      type: integer
                    low:
                      type: integer
                byAgent:
                  type: object
                  additionalProperties:
                    type: integer
            nextTask:
              type: object
              properties:
                id:
                  type: string
                type:
                  type: string
                priority:
                  type: string
                assignedAgent:
                  type: string
                estimatedDuration:
                  type: integer
            config:
              type: object
              properties:
                maxSize:
                  type: integer
                maxAttempts:
                  type: integer
                defaultPriority:
                  type: string
    EnqueueTaskRequest:
      type: object
      required: [id, type, payload]
      properties:
        id:
          type: string
        type:
          type: string
          description: 任务类型
        priority:
          type: string
          enum: [urgent, high, normal, low]
          default: normal
        payload:
          type: object
          description: 任务数据
        requiredCapabilities:
          type: array
          items:
            type: string
        preferredAgent:
          type: string
        maxAttempts:
          type: integer
          minimum: 1
          maximum: 10
          default: 3
        estimatedDuration:
          type: integer
        dependencies:
          type: array
          items:
            type: string
        metadata:
          type: object
    EnqueueTaskResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            taskId:
              type: string
            queuePosition:
              type: integer
            assignedAgent:
              type: string
              nullable: true
            estimatedStart:
              type: string
              format: date-time
            scheduling:
              type: object
              properties:
                candidates:
                  type: array
                  items:
                    type: object
                    properties:
                      agentId:
                        type: string
                      score:
                        type: number
                      currentLoad:
                        type: integer
                      availableAt:
                        type: string
                        format: date-time
                reasoning:
                  type: string

    # ===== JSON-RPC Schema =====
...
paths:
...
  /a2a/queue:
    get:
      tags: [Agents, Tasks]
      summary: 获取任务队列状态
      operationId: getQueueStatus
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, scheduled, in_progress, completed, failed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [urgent, high, normal, low]
        - name: agentId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueStatusResponse'

    post:
      tags: [Agents, Tasks]
      summary: 添加任务到队列
      operationId: enqueueTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnqueueTaskRequest'
      responses:
        '201':
          description: 任务已入队
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnqueueTaskResponse'
        '400':
          description: 参数验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Agents, Tasks]
      summary: 清空任务队列
      operationId: clearQueue
      security:
        - bearerAuth: []
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
            enum: [urgent, high, normal, low]
      responses:
        '200':
          description: 清空成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      removed:
                        type: integer
                      remaining:
                        type: integer
...